bam/record/quality_scores: Avoid iterator adapter for BAM to SAM conversion
For some reason, collecting into a `Result<Vec<_>, _>` makes this spend
a lot of time in `alloc::vec::Vec<T>::extend_desugared`. It may have to
do with the way `size_hint` is defined on
`core::iter::adapters::ResultShunt<I, E>` and used when processing the
results, which causes excessive allocations.