sam/alignment_reader: Changes `read_alignment_header` to return `String`
The rationale for this change is simple: similar to the other
`read_header` methods, returning a `String` allows the user to
parse the `sam::Header` _after_ applying corrections to raw
header text. This is desirable in cases where the header may not
conform to the specification, so preprocessing of the header
is required before parsing.