sam/alignment/iter: Add column depth iterator
This is a simple pileup iterator that simply calculate read depth. It
piles records over an adaptive window on the reference sequence and is
optimized for low latency, i.e., it emits columns immediately after they
are no longer affected by future records. This implies that it only
works with coordinate-sorted data.

Thanks to @sstadick for the initial implementation and inspiration in
#14.