util/alignment/reader: Fix format detection when the first BGZF block is incomplete
The BGZF decoder requires full blocks when decoding. When the first BGZF
block is incomplete, it is not possible to get the start of the
uncompressed output, which would return `ErrorKind::UnexpectedEof`.
However, the error was silently being discarded.

The BGZF decoder is now changed to a streaming multistream gzip decoder.

Fixes #179.