sam/header/programs: Add method to attach program to program chains
If the program is the first program in the graph, this is similar to
calling `IndexMap::insert` on the inner graph. If no previous program is
set, this attaches the program to all program chains using leaf programs
as the given program's previous program.

Thanks to @jamestwebber for initially exploring this in #245.

Closes #245.